Changes to VXM CMD convention. We use Im Mx to denote the motor being indexed, and the number of steps moved

+++ print_ctrl/compiler/laser/getStatus.m
@@ -1,6 +1,6 @@
 % M201: Get Status
 % Does targeting laser for now
 function laserCMD = getStatus(command)
-    laserCMD = [0x1B, 0x01, command, 0x0D, 0x2A];
+    laserCMD = [0x1B, 0x01, uint8(command), 0x0D, 0x2A];
     % compose("0x%02X", getStatus) To view
 end
